+++ .github/workflows/release.yml
@@ -1,0 +1,34 @@
+name: Build GNNWorker
+
+on:
+  push:
+    tags:
+      - "v*.*.*"
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout repo
+        uses: actions/checkout@v4
+
+      - name: Install Julia
+        uses: julia-actions/setup-julia@v1
+        with:
+          version: '1.11'
+
+      - name: Build executable
+        run: bash GNN/scripts/build.sh
+
+      - name: Zip build
+        run: |
+          cd GNN/build
+          zip -r GNNWorker.zip GNNWorker
+
+      - name: Upload release
+        uses: softprops/action-gh-release@v2
+        with:
+          files: GNN/build/GNNWorker.zip
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
