+++ GNN/Project.toml
@@ -4,68 +4,46 @@ authors = ["Elian H. Thiele-Evans <60372411+ElianHugh@users.noreply.github.com>"
 version = "0.1.0"
 
 [deps]
-Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
+ArgParse = "c7e460c6-2fb9-53a9-8c5b-16f535851c63"
 BSON = "fbb218c0-5317-5bc6-957e-2ee96dd4b1f0"
-CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
-CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 Clustering = "aaaa29a8-35af-508c-8bc3-b662a17a0fe5"
-Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
-DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Distances = "b4f34e82-e78d-54a5-968a-f98e89d6e8f7"
 Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
-Functors = "d9f16b24-f501-4c13-a1f2-28368ffc5196"
 GNNGraphs = "aed8fd31-079b-4b5a-b342-a13352159b8c"
-GraphMakie = "1ecd5474-83a3-4783-bb4f-06765db800d2"
 GraphNeuralNetworks = "cffab07f-9bc2-4db1-8861-388f63bf7694"
-GraphPlot = "a2cc645c-3eea-5389-862e-a155d0052231"
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
-HTTP = "cd3eb016-35fb-5094-929b-558a96fad6f3"
 JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
 Leiden = "8876827e-33c4-4373-8179-6e3204d93767"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-Mongoc = "4fe8b98c-fc19-5c23-8ec2-168ff83495f2"
-MultivariateStats = "6f286f6a-111f-5878-ab1e-185364afe411"
-NetworkLayout = "46757867-2c16-5918-afeb-47bfcb05e46a"
-OpenCV = "f878e3a2-a245-4720-8660-60795d644f2a"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+PackageCompiler = "9b87118b-4619-50d2-8e1e-99f35a4d4d9d"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
-UMAP = "c4f8c510-2410-5be4-91d7-4fbaeb39457e"
+StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 XLSX = "fdbf4ff8-1666-58a4-91e7-1b58723a45e0"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 igraph_jll = "4d841945-e0c2-5143-9479-18e4f0102556"
 leiden_jll = "b3d7b256-84d9-5f32-9065-d6f1a433f499"
 
 [compat]
-Adapt = "4.3.0"
+ArgParse = "1.2.0"
 BSON = "0.3.9"
-CUDA = "5.7.3"
-CairoMakie = "0.13.4"
 Clustering = "0.15.8"
-Colors = "0.12.11"
-DataFrames = "1.7.0"
 Distances = "0.10.12"
 Flux = "0.16.3"
-Functors = "0.5.2"
 GNNGraphs = "1.4.2"
-GraphMakie = "0.5.14"
 GraphNeuralNetworks = "1.0.0"
-GraphPlot = "0.6.1"
 Graphs = "1.12.1"
-HTTP = "1.10.16"
 JSON3 = "1.14.2"
 Leiden = "0.1.0"
 LinearAlgebra = "1.11.0"
-Mongoc = "0.9.2"
-MultivariateStats = "0.10.3"
-NetworkLayout = "0.4.10"
-OpenCV = "4.6.1"
-Plots = "1.40.13"
+PackageCompiler = "2.2.0"
 Random = "1.11.0"
 Reexport = "1.2.2"
 Statistics = "1.11.1"
-UMAP = "0.1.11"
+StatsBase = "0.34.5"
+Test = "1.11.0"
 XLSX = "0.10.4"
 Zygote = "0.7.6"
 igraph_jll = "0.9.10"
