--- GNN/output/logs/model_summary.md
@@ -1,57 +1,0 @@
-# Baseline Graph Summary
-
-## Cluster Quality
-
-- quality:
-  - modularity = 0.49019676169243825
-  - conductance = 0.38038358852661663
-- n_clusters = 9
-
-## Intra-cluster Composite Rate
-
-- Negative: 0.06578947368421052
-- Positive: 0.6492838780756518
-
-## Per-view Cluster Rate
-
-- influence = 0.4963768115942029
-- friendship = 0.7091531223267751
-- more_time = 0.6829268292682927
-- advice = 0.7220708446866485
-- disrespect = 0.06578947368421052
-- feedback = 0.7887323943661971
-- affiliation = 0.3090277777777778
-
-# Model Summary
-
-## Hyperparameters
-
-- γ = 0.01
-- λ = 10.0
-- τ = 1.0
-- epochs = 500
-
-## Cluster Quality
-
-- var_norm = 15040.544
-- quality:
-  - modularity = 0.19200166036943986
-  - silhouettes = 0.17628092
-  - conductance = 0.7734527369966656
-- n_clusters = 4
-- mean_norm = 105.51402
-
-## Intra-cluster Composite Rate
-
-- Negative: 0.17105263157894737
-- Positive: 0.612559676827029
-
-## Per-view Cluster Rate
-
-- influence = 0.44565217391304346
-- friendship = 0.6467065868263473
-- more_time = 0.7170731707317073
-- advice = 0.6893732970027248
-- disrespect = 0.17105263157894737
-- feedback = 0.676056338028169
-- affiliation = 0.3402777777777778
