+++ GNN/src/GNNProject.jl
@@ -1,12 +1,13 @@
 module GNNProject
     using Reexport
 
-    include("types.jl")
-    include("data_loader.jl")
-    include("loss.jl")
-    include("model_evaluation.jl")
-    include("training.jl")
-    include("plotting.jl")
+    include("Types.jl")
+    include("DataLoader.jl")
+    include("Loss.jl")
+    include("ModelEvaluation.jl")
+    include("Training.jl")
+    include("Plotting.jl")
+    include("ClusterWorker.jl")
 
     @reexport using .Types
     @reexport using .DataLoader
@@ -14,5 +15,6 @@ module GNNProject
     @reexport using .ModelEvaluation
     @reexport using .ModelTraining
     @reexport using .Plotting
+    @reexport using .ClusterWorker
 
 end
